make camera a feature
